pyproject.toml
[tool.poetry]
name = "mm-detect"
version = "1.0.0"
description = "The First Multimodal Data Contamination Detection Framework"
authors = [
    "Sicheng Lai <sicheng.lai@example.com>",
]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/MLLM-Data-Contamination/MM-Detect.git"
repository = "https://github.com/MLLM-Data-Contamination/MM-Detect.git"
documentation = "https://github.com/MLLM-Data-Contamination/MM-Detect.git"
keywords = ["multimodal", "data-contamination", "llm", "mllm", "detection"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

[tool.poetry.dependencies]
python = "^3.10"
torch = "2.1.2"
datasets = "2.17.1"
evaluate = "0.4.1"
huggingface-hub = "0.20.3"
matplotlib = "3.8.3"
nltk = "3.8.1"
numpy = "1.26.4"
openai = "1.35.7"
pandas = "2.2.1"
rouge-score = "0.1.2"
scipy = "1.12.0"
sentence-transformers = "3.0.1"
sentencepiece = "0.2.0"
transformers = "4.37.2"
triton = "2.1.0"
accelerate = "1.3.0"
googletrans = "3.1.0a0"
httpx = "0.27.2"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.0.0"
flake8 = "^6.0.0"
isort = "^5.12.0"
mypy = "^1.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py310"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["mm_detect"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true